cheque form updated

+++ AUTOPARC/Pages/Vehicule/Details.cshtml.cs
@@ -50,9 +50,9 @@ namespace AUTOPARC.Pages.Vehicule
         public ImageVehicule ImageVehicule { get; set; }
 
         public bool check_vehicule_exception, check_vehicule_matricule, check_all_checkboxes;
-        public bool check_cheque_exception, check_numero_cheque_existance, check_cheque_date;
-        public bool check_virement_exception;
-        public bool check_credit_exception, check_credit_montantParMoi, check_credit_date;
+        public bool check_cheque_exception, check_numero_cheque_existance, check_cheque_date, check_cheque_payee;
+        public bool check_virement_exception, check_virement_payee;
+        public bool check_credit_exception, check_credit_montantParMoi, check_credit_date, check_credit_payee;
 
 
 
@@ -84,35 +84,22 @@ namespace AUTOPARC.Pages.Vehicule
             bool isVirement = Request.Form.TryGetValue("Virement", out var virementValue);
             bool isCredit = Request.Form.TryGetValue("Credit", out var creditValue);
 
-            if (!isEspece && !isCheque && !isVirement && !isCredit)
-            {
-                check_all_checkboxes = true;
-                await OnGet(Vehicules.MarqueId);
-                return Page();
-            }
+            if (!isCheque)
+                await RemoveChequeModelStateEntriesAsync();
 
+            if (!isVirement)
+                await RemoveVirementModelStateEntriesAsync();
 
-            //if (await Check_ChequePayee())
-            //    return Page();
-
-
-            //if (Vehicules.PrixAchat < Vehicules.MontantPayeeEspece)
-            //{
-            //    check_Prix_MontantPayee = true;
-            //    await OnGet(Vehicules.Id);
-            //    return Page();
-            //}
-
-
-            //var modePaiement = await _db.ModePaiments.Where(m => m.Id == Vehicules.ModePayementId).Select(m => m.Mode).SingleOrDefaultAsync();
-            //if (modePaiement == "Espece")
-            //    await RemoveChequeModelStateEntriesAsync();
-            //else if (!await TryUpdateCheque())
-            //    return Page();
-
+            if (!isCredit)
+                await RemoveCreditModelStateEntriesAsync();
 
             try
             {
+
+
+
+
+
                 var vehicule = await _db.Vehicules.FindAsync(Vehicules.Id);
                 vehicule.Matricule = Vehicules.Matricule;
                 vehicule.CarburantId = Vehicules.CarburantId;
@@ -122,8 +109,6 @@ namespace AUTOPARC.Pages.Vehicule
                 vehicule.EtatVehiculeId = Vehicules.EtatVehiculeId;
                 vehicule.FrsId = Vehicules.FrsId;
                 vehicule.PrixAchat = Vehicules.PrixAchat;
-                vehicule.MontantPayeeEspece = Vehicules.MontantPayeeEspece;
-                //vehicule.ModePayementId = Vehicules.ModePayementId;
                 vehicule.DateAchat = Vehicules.DateAchat;
                 vehicule.DateMisEnCirculation = Vehicules.DateMisEnCirculation;
                 vehicule.Kilometrage = Vehicules.Kilometrage;
@@ -142,12 +127,29 @@ namespace AUTOPARC.Pages.Vehicule
                 vehicule.Hauteur = Vehicules.Hauteur;
                 vehicule.VolumeCoffre = Vehicules.VolumeCoffre;
 
+
+                if (isCheque && !await UpdateChequeAsync())
+                    return Page();
+
+                if (isVirement && !await UpdateVirementAsync())
+                    return Page();
+
+                if (isCredit && !await UpdateCreditAsync())
+                    return Page();
+
+
+                vehicule.MontantPayeeTotal -= vehicule.MontantPayeeEspece;
+                vehicule.MontantPayeeEspece = Vehicules.MontantPayeeEspece;
+                vehicule.MontantPayeeCheque = Vehicules.MontantPayeeCheque;
+                vehicule.MontantPayeeVirement = Vehicules.MontantPayeeVirement;
+                vehicule.MontantPayeeTotal += vehicule.MontantPayeeEspece;
+
                 await _db.SaveChangesAsync();
                 return RedirectToPage("/Vehicule/Index");
             }
             catch (Exception)
             {
-                //check_exception = true;
+                check_vehicule_exception = true;
                 await OnGet(Vehicules.Id);
                 return Page();
             }
@@ -161,7 +163,14 @@ namespace AUTOPARC.Pages.Vehicule
             //if (await Check_ChequePayee())
             //    return Page();
 
-            await TryDeleteCheck();
+            if (!await TryDeleteCheck())
+                return Page();
+
+            if (!await TryDeleteVirement())
+                return Page();
+
+            if (!await TryDeleteCredit())
+                return Page();
 
             _db.Vehicules.Remove(Vehicules);
             await _db.SaveChangesAsync();
@@ -172,28 +181,20 @@ namespace AUTOPARC.Pages.Vehicule
 
 
 
-
-        private async Task<bool> UpdateCheque()
+        private async Task<bool> UpdateChequeAsync()
         {
             if (string.IsNullOrEmpty(Cheques.Numero) || string.IsNullOrEmpty(Cheques.AuNomDe) ||
                 Cheques.DateReglement == null || Cheques.DateEcheance == null ||
-                Cheques.Montant == 0 || Cheques.BanqueId == 0)
+                Cheques.Montant <= 0 || Cheques.BanqueId == 0)
             {
                 check_cheque_exception = true;
                 await OnGet(Vehicules.Id);
                 return false;
             }
 
-            if (Vehicules.PrixAchat != (Vehicules.MontantPayeeEspece + Cheques.Montant))
-            {
-                //check_montant = true;
-                await OnGet(Vehicules.Id);
-                return false;
-            }
-
             if (Cheques.DateReglement > Cheques.DateEcheance)
             {
-                //check_date = true;
+                check_cheque_date = true;
                 await OnGet(Vehicules.Id);
                 return false;
             }
@@ -211,18 +212,15 @@ namespace AUTOPARC.Pages.Vehicule
                     cheque.DateReglement = Cheques.DateReglement;
                     cheque.DateEcheance = Cheques.DateEcheance;
                     cheque.DateValeur = Cheques.DateValeur;
+                    Vehicules.MontantPayeeCheque = Cheques.Montant;
 
-                    await _db.SaveChangesAsync();
                     return true;
                 }
-                else
-                {
-                    Cheques.ActionNum = Vehicules.Num;
-                    Cheques.Etat = "impayé";
-                    await _db.Cheques.AddAsync(Cheques);
-                    await _db.SaveChangesAsync();
-                    return true;
-                }
+
+                Vehicules.MontantPayeeCheque = Cheques.Montant;
+                Cheques.ActionNum = Vehicules.Num;
+                await _db.Cheques.AddAsync(Cheques);
+                return true;
             }
             catch (DbUpdateException ex) when (ex.InnerException is MySqlException mySqlEx)
             {
@@ -246,101 +244,25 @@ namespace AUTOPARC.Pages.Vehicule
 
 
 
-
-
-
-
-
-        private async Task TryDeleteCheck()
-        {
-            var cheque = await _db.Cheques.Where(chq => chq.Action == "Vehicule" && chq.ActionNum == Vehicules.Num).FirstOrDefaultAsync();
-            if (cheque != null)
-            {
-                _db.Cheques.Remove(cheque);
-                await _db.SaveChangesAsync();
-            }
-        }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-        private async Task<bool> InsertChequeAsync()
-        {
-            if (string.IsNullOrEmpty(Cheques.Numero) || string.IsNullOrEmpty(Cheques.AuNomDe) ||
-                Cheques.DateReglement == null || Cheques.DateEcheance == null ||
-                Cheques.Montant == 0 || Cheques.BanqueId == 0)
-            {
-                check_cheque_exception = true;
-                await OnGet(Vehicules.MarqueId);
-                return false;
-            }
-
-            if (Cheques.DateReglement > Cheques.DateEcheance)
-            {
-                check_cheque_date = true;
-                await OnGet(Vehicules.MarqueId);
-                return false;
-            }
-
-            try
-            {
-                Vehicules.MontantPayeeCheque = Cheques.Montant;
-                Cheques.ActionNum = Vehicules.Num;
-                await _db.Cheques.AddAsync(Cheques);
-                return true;
-            }
-            catch (DbUpdateException ex) when (ex.InnerException is MySqlException mySqlEx)
-            {
-                if (mySqlEx.Message.Contains("Numero"))
-                {
-                    check_numero_cheque_existance = true;
-                    await OnGet(Vehicules.MarqueId);
-                }
-                return false;
-            }
-            catch
-            {
-                check_cheque_exception = true;
-                await OnGet(Vehicules.MarqueId);
-                return false;
-            }
-        }
-
-
-
-
-        private async Task<bool> InsertVirementAsync()
+        private async Task<bool> UpdateVirementAsync()
         {
             if (string.IsNullOrEmpty(Virements.Rib) || string.IsNullOrEmpty(Virements.AuNomDe) ||
-                Virements.Montant == 0 || Virements.BanqueId == 0 ||
-                 Virements.DateVirement == null)
+                Virements.DateVirement == null || Virements.Montant <= 0 || Virements.BanqueId == 0)
             {
-                check_cheque_exception = true;
-                await OnGet(Vehicules.MarqueId);
+                check_virement_exception = true;
+                await OnGet(Vehicules.Id);
                 return false;
             }
 
             try
             {
+                var virement = await _db.Virements.Where(v => v.Action == "Vehicule" && v.ActionNum == Vehicules.Num).FirstOrDefaultAsync();
+
+                if (virement != null)
+                    return true;
+
                 Vehicules.MontantPayeeVirement = Virements.Montant;
+                Vehicules.MontantPayeeTotal += Virements.Montant;
                 Virements.ActionNum = Vehicules.Num;
                 await _db.Virements.AddAsync(Virements);
                 return true;
@@ -348,7 +270,7 @@ namespace AUTOPARC.Pages.Vehicule
             catch
             {
                 check_virement_exception = true;
-                await OnGet(Vehicules.MarqueId);
+                await OnGet(Vehicules.Id);
                 return false;
             }
         }
@@ -356,44 +278,96 @@ namespace AUTOPARC.Pages.Vehicule
 
 
 
-        private async Task<bool> InsertCreditAsync()
+
+
+        private async Task<bool> UpdateCreditAsync()
         {
-            if (Credits.Montant == 0 || Credits.Mensualite == 0 || Credits.BanqueId == 0 ||
+            if (Credits.Montant <= 0 || Credits.Mensualite <= 0 || Credits.BanqueId == 0 ||
                 Credits.DateDebut == null || Credits.DateFin == null)
             {
                 check_credit_exception = true;
-                await OnGet(Vehicules.MarqueId);
-                return false;
-            }
-
-            if (Credits.Montant < Credits.Mensualite)
-            {
-                check_credit_montantParMoi = true;
-                await OnGet(Vehicules.MarqueId);
-                return false;
-            }
-
-            if (Credits.DateDebut > Credits.DateFin)
-            {
-                check_credit_date = true;
-                await OnGet(Vehicules.MarqueId);
+                await OnGet(Vehicules.Id);
                 return false;
             }
 
             try
             {
+                var credit = await _db.Credits.Where(c => c.Action == "Vehicule" && c.ActionNum == Vehicules.Num).FirstOrDefaultAsync();
+
+                if (credit != null)
+                {
+                    credit.Mensualite = Credits.Mensualite;
+                    credit.DateFin = Credits.DateFin;
+                    return true;
+                }
+
                 Vehicules.MontantPayeeCredit = Credits.Montant;
                 Credits.ActionNum = Vehicules.Num;
+                Credits.Etat = "impayé";
                 await _db.Credits.AddAsync(Credits);
                 return true;
             }
             catch
             {
-                check_virement_exception = true;
-                await OnGet(Vehicules.MarqueId);
+                check_credit_exception = true;
+                await OnGet(Vehicules.Id);
                 return false;
             }
         }
+
+
+
+
+
+        private async Task<bool> TryDeleteCheck()
+        {
+            var cheque = await _db.Cheques.Where(chq => chq.Action == "Vehicule" && chq.ActionNum == Vehicules.Num).FirstOrDefaultAsync();
+            if (cheque != null && cheque.Etat != "payé")
+            {
+                _db.Cheques.Remove(cheque);
+                return true;
+            }
+            else if(cheque != null && cheque.Etat == "payé")
+            {
+                check_cheque_payee = true;
+                await OnGet(Vehicules.Id);
+                return false;
+            }
+            return true;
+        }
+
+        private async Task<bool> TryDeleteVirement()
+        {
+            var virement = await _db.Virements.Where(v => v.Action == "Vehicule" && v.ActionNum == Vehicules.Num).FirstOrDefaultAsync();
+            if (virement != null && virement.Etat != "payé")
+            {
+                _db.Virements.Remove(virement);
+                return true;
+            }
+            else if (virement != null && virement.Etat == "payé")
+            {
+                check_virement_payee = true;
+                await OnGet(Vehicules.Id);
+                return false;
+            }
+            return true;
+        }
+
+        private async Task<bool> TryDeleteCredit()
+        {
+            var credit = await _db.Credits.Where(c => c.Action == "Vehicule" && c.ActionNum == Vehicules.Num).FirstOrDefaultAsync();
+
+            if (credit == null)
+                return true;
+
+            check_credit_payee = true;
+            await OnGet(Vehicules.Id);
+            return false;
+        }
+
+
+
+
